requirements.txt
tensorflow
numpy
matplotlib
pillow
pandas
scikit-learn